Stealth Shader Update + Packaging Cleanup
Since the Flat Lit + MMD spheres is starting to grow into different branches, I decided to package them all under one header for easier use.  On top of that I revisited the stealth effect shader and updated it with all the various tweaks the other ones have received over the past while and FINALLY made an editor for it that doesn't explode.
Also added an toggle for detail normals to use two different types of specular calculations based on popular demand.