Ok I Officially Can't Call it Flat Lit Anymore
Real-time lighting support is now here!  It's rough and tumble, and will probably need a good half dozen little updates and changes but it works.
To a point.

Now in any real-time lit situation with shadows enabled, there will be self shadowing based upon the toon ramp of the model.  There are some plans on also updating the sphere mask of the model so that it reacts to real time lighting as well, but that's for the future.
Also, Transparency works now with an opacity slider.  At 0 it's the base alpha of the texture, at 1 it's 100% transparent.  At full transparency things can be a little wonky so just be aware.
Also a detail normals error should be all fixed up with strange alternate eye normals.

I know the files are still called Flat Lit by the way.
I'll change that eventually.

PS:
Changed the names of Transparent and Transparent Special to Transparent Lit and Transparent Unlit, Lit is on queue Transparent-2 and can receive lighting so use it on semi-transparent clothes in cutout mode to self-shadow.  Unlit is on queue Transparent-1 and should mainly be used for effects like blush.