1.82 Update No. 2
Added in a Sprite Sheet shader, it's super prototype-y and ask me about how it works before just going in and playing with it.
The big updates here are a fix for the Transparent Lit shader and the Shadows.
With the shadows no longer will transparent cutout sections block shadows leading to some pretty gnarly harsh cuts.