Version 2.0
It's been a while but I've finally finished my big overhaul of the backend systems.
What does that mean for you?
Absolutely nothing!
But it means for me that I'll have a lot less headaches updating and adding on new functionality for shaders.
I have touched up on a couple things:
-Cleaned out some older, obsolete versions
-Updated the Stealth shader for the first time in forever (still be careful, haven't testing it in-game)
-Updated Fader Shader with another variant to apply a normal map specifically to the 2nd matcap as well as functionality so that it doesn't overwrite itself.